// packages/axelar/src/codegen/axelar/tss/tofnd/v1beta1/tofnd.d.ts
export interface KeygenOutput {
    /** pub_key; common for all parties */
    pubKey: Uint8Array;
    /** recover info of all parties' shares; common for all parties */
    groupRecoverInfo: Uint8Array;
    /** private recover info of this party's shares; unique for each party */
    privateRecoverInfo: Uint8Array;
}
export interface KeygenOutputProtoMsg {
    typeUrl: "/axelar.tss.tofnd.v1beta1.KeygenOutput";
    value: Uint8Array;
}
/** Keygen's success response */
export interface KeygenOutputAmino {
    /** pub_key; common for all parties */
    pub_key?: string;
    /** recover info of all parties' shares; common for all parties */
    group_recover_info?: string;
    /** private recover info of this party's shares; unique for each party */
    private_recover_info?: string;
}
export interface KeygenOutputAminoMsg {
    type: "/axelar.tss.tofnd.v1beta1.KeygenOutput";
    value: KeygenOutputAmino;
}
// ...
export interface MessageOut_KeygenResult {
    /** Success response */
    data?: KeygenOutput;
    /** Faiilure response */
    criminals?: MessageOut_CriminalList;
}
export interface MessageOut_KeygenResultProtoMsg {
    typeUrl: "/axelar.tss.tofnd.v1beta1.KeygenResult";
    value: Uint8Array;
}
/** Keygen's response types */
export interface MessageOut_KeygenResultAmino {
    /** Success response */
    data?: KeygenOutputAmino;
    /** Faiilure response */
    criminals?: MessageOut_CriminalListAmino;
}
export interface MessageOut_KeygenResultAminoMsg {
    type: "/axelar.tss.tofnd.v1beta1.KeygenResult";
    value: MessageOut_KeygenResultAmino;
}
/** Keygen's response types */
export interface MessageOut_KeygenResultSDKType {
    data?: KeygenOutputSDKType;
    criminals?: MessageOut_CriminalListSDKType;
}
/** Sign's response types */
export interface MessageOut_SignResult {
    /** Success response */
    signature?: Uint8Array;
    /** Failure response */
    criminals?: MessageOut_CriminalList;
}
export interface MessageOut_SignResultProtoMsg {
    typeUrl: "/axelar.tss.tofnd.v1beta1.SignResult";
    value: Uint8Array;
}
/** Sign's response types */
export interface MessageOut_SignResultAmino {
    /** Success response */
    signature?: string;
    /** Failure response */
    criminals?: MessageOut_CriminalListAmino;
}
export interface MessageOut_SignResultAminoMsg {
    type: "/axelar.tss.tofnd.v1beta1.SignResult";
    value: MessageOut_SignResultAmino;
}
/** Sign's response types */
export interface MessageOut_SignResultSDKType {
    signature?: Uint8Array;
    criminals?: MessageOut_CriminalListSDKType;
}
/** Keygen/Sign failure response message */
export interface MessageOut_CriminalList {
    criminals: MessageOut_CriminalList_Criminal[];
}
export interface MessageOut_CriminalListProtoMsg {
    typeUrl: "/axelar.tss.tofnd.v1beta1.CriminalList";
    value: Uint8Array;
}
/** Keygen/Sign failure response message */
export interface MessageOut_CriminalListAmino {
    criminals?: MessageOut_CriminalList_CriminalAmino[];
}
export interface MessageOut_CriminalListAminoMsg {
    type: "/axelar.tss.tofnd.v1beta1.CriminalList";
    value: MessageOut_CriminalListAmino;
}
/** Keygen/Sign failure response message */
export interface MessageOut_CriminalListSDKType {
    criminals: MessageOut_CriminalList_CriminalSDKType[];
}
export interface MessageOut_CriminalList_Criminal {
    partyUid: string;
    crimeType: MessageOut_CriminalList_Criminal_CrimeType;
}
// ...
export interface TrafficOut {
    toPartyUid: string;
    payload: Uint8Array;
    isBroadcast: boolean;
}
export interface TrafficOutProtoMsg {
    typeUrl: "/axelar.tss.tofnd.v1beta1.TrafficOut";
    value: Uint8Array;
}
export interface TrafficOutAmino {
    to_party_uid?: string;
    payload?: string;
    is_broadcast?: boolean;
}
export interface TrafficOutAminoMsg {
    type: "/axelar.tss.tofnd.v1beta1.TrafficOut";
    value: TrafficOutAmino;
}
export interface TrafficOutSDKType {
    to_party_uid: string;
    payload: Uint8Array;
    is_broadcast: boolean;
}
// ...
export declare const KeygenOutput: {
    typeUrl: string;
    encode(message: KeygenOutput, writer?: BinaryWriter): BinaryWriter;
    fromJSON(object: any): KeygenOutput;
    fromPartial(object: Partial<KeygenOutput>): KeygenOutput;
    fromAmino(object: KeygenOutputAmino): KeygenOutput;
    toAmino(message: KeygenOutput): KeygenOutputAmino;
    fromAminoMsg(object: KeygenOutputAminoMsg): KeygenOutput;
    fromProtoMsg(message: KeygenOutputProtoMsg): KeygenOutput;
    toProto(message: KeygenOutput): Uint8Array;
    toProtoMsg(message: KeygenOutput): KeygenOutputProtoMsg;
};
// ...
export declare const MessageOut_KeygenResult: {
    typeUrl: string;
    encode(message: MessageOut_KeygenResult, writer?: BinaryWriter): BinaryWriter;
    fromJSON(object: any): MessageOut_KeygenResult;
    fromPartial(object: Partial<MessageOut_KeygenResult>): MessageOut_KeygenResult;
    fromAmino(object: MessageOut_KeygenResultAmino): MessageOut_KeygenResult;
    toAmino(message: MessageOut_KeygenResult): MessageOut_KeygenResultAmino;
    fromAminoMsg(object: MessageOut_KeygenResultAminoMsg): MessageOut_KeygenResult;
    fromProtoMsg(message: MessageOut_KeygenResultProtoMsg): MessageOut_KeygenResult;
    toProto(message: MessageOut_KeygenResult): Uint8Array;
    toProtoMsg(message: MessageOut_KeygenResult): MessageOut_KeygenResultProtoMsg;
};
export declare const MessageOut_SignResult: {
    typeUrl: string;
    encode(message: MessageOut_SignResult, writer?: BinaryWriter): BinaryWriter;
    fromJSON(object: any): MessageOut_SignResult;
    fromPartial(object: Partial<MessageOut_SignResult>): MessageOut_SignResult;
    fromAmino(object: MessageOut_SignResultAmino): MessageOut_SignResult;
    toAmino(message: MessageOut_SignResult): MessageOut_SignResultAmino;
    fromAminoMsg(object: MessageOut_SignResultAminoMsg): MessageOut_SignResult;
    fromProtoMsg(message: MessageOut_SignResultProtoMsg): MessageOut_SignResult;
    toProto(message: MessageOut_SignResult): Uint8Array;
    toProtoMsg(message: MessageOut_SignResult): MessageOut_SignResultProtoMsg;
};
export declare const MessageOut_CriminalList: {
    typeUrl: string;
    encode(message: MessageOut_CriminalList, writer?: BinaryWriter): BinaryWriter;
    fromJSON(object: any): MessageOut_CriminalList;
    fromPartial(object: Partial<MessageOut_CriminalList>): MessageOut_CriminalList;
    fromAmino(object: MessageOut_CriminalListAmino): MessageOut_CriminalList;
    toAmino(message: MessageOut_CriminalList): MessageOut_CriminalListAmino;
    fromAminoMsg(object: MessageOut_CriminalListAminoMsg): MessageOut_CriminalList;
    fromProtoMsg(message: MessageOut_CriminalListProtoMsg): MessageOut_CriminalList;
    toProto(message: MessageOut_CriminalList): Uint8Array;
    toProtoMsg(message: MessageOut_CriminalList): MessageOut_CriminalListProtoMsg;
};
export declare const MessageOut_CriminalList_Criminal: {
    typeUrl: string;
    encode(message: MessageOut_CriminalList_Criminal, writer?: BinaryWriter): BinaryWriter;
    fromJSON(object: any): MessageOut_CriminalList_Criminal;
    fromPartial(object: Partial<MessageOut_CriminalList_Criminal>): MessageOut_CriminalList_Criminal;
    fromAmino(object: MessageOut_CriminalList_CriminalAmino): MessageOut_CriminalList_Criminal;
    toAmino(message: MessageOut_CriminalList_Criminal): MessageOut_CriminalList_CriminalAmino;
    fromAminoMsg(object: MessageOut_CriminalList_CriminalAminoMsg): MessageOut_CriminalList_Criminal;
    fromProtoMsg(message: MessageOut_CriminalList_CriminalProtoMsg): MessageOut_CriminalList_Criminal;
    toProto(message: MessageOut_CriminalList_Criminal): Uint8Array;
    toProtoMsg(message: MessageOut_CriminalList_Criminal): MessageOut_CriminalList_CriminalProtoMsg;
};
// ...
export declare const TrafficOut: {
    typeUrl: string;
    encode(message: TrafficOut, writer?: BinaryWriter): BinaryWriter;
    fromJSON(object: any): TrafficOut;
    fromPartial(object: Partial<TrafficOut>): TrafficOut;
    fromAmino(object: TrafficOutAmino): TrafficOut;
    toAmino(message: TrafficOut): TrafficOutAmino;
    fromAminoMsg(object: TrafficOutAminoMsg): TrafficOut;
    fromProtoMsg(message: TrafficOutProtoMsg): TrafficOut;
    toProto(message: TrafficOut): Uint8Array;
    toProtoMsg(message: TrafficOut): TrafficOutProtoMsg;
};